change seed to int64_t as POSIX.1-2024 requires time_t to be 64 bits wide

+++ src/msolve/main.c
@@ -179,7 +179,7 @@ static void getoptions(
         int32_t *refine,
         int32_t *isolate,
         int32_t *generate_pbm_files,
-	int32_t *seed,
+	int64_t *seed,
         int32_t *info_level,
         files_gb *files){
   int opt, errflag = 0, fflag = 1;
@@ -386,7 +386,7 @@ int main(int argc, char **argv){
     int32_t precision             = 64;
     int32_t refine                = 0; /* not used at the moment */
     int32_t isolate               = 0; /* not used at the moment */
-    int32_t seed                  = 0;
+    int64_t seed                  = 0;
 
     files_gb *files = malloc(sizeof(files_gb));
     if(files == NULL) exit(1);
@@ -408,7 +408,7 @@ int main(int argc, char **argv){
     srand(seed);
     if (info_level) {
       fprintf (stdout,"Initial seed for pseudo-random number generator ");
-      fprintf (stdout,"is %d\n",seed);
+      fprintf (stdout,"is %ld\n",seed);
     }
 
     FILE *fh  = fopen(files->in_file, "r");
